wip/weekly bundle compiles now
